refactor: InputParser.java is better defined as a view
