+++ src/main/java/com/om1cael/view/InputParser.java
@@ -1,4 +1,4 @@
-package com.om1cael.controller;
+package com.om1cael.view;
 
 import java.util.Scanner;
